+++ lib/hanami/interactor.rb
@@ -1,6 +1,7 @@
 require 'hanami/utils/basic_object'
 require 'hanami/utils/class_attribute'
 require 'hanami/utils/hash'
+require 'hanami/utils/deprecation'
 
 module Hanami
   # Hanami Interactor
@@ -20,6 +21,7 @@ module Hanami
       METHODS = ::Hash[initialize:  true,
                        success?:    true,
                        successful?: true,
+                       failure?:    true,
                        failing?:    true,
                        fail!:       true,
                        prepare!:    true,
@@ -56,8 +58,20 @@ module Hanami
       #
       # @return [TrueClass,FalseClass] the result of the check
       #
-      # @since 0.8.1
+      # @since 0.9.2
+      #
+      # @deprecated Use {#failure?} instead
       def failing?
+        Utils::Deprecation.new("`Hanami::Interactor::Result#failing?' is deprecated, please use `Hanami::Interactor::Result#failure?'")
+        failure?
+      end
+
+      # Check if the current status is not successful
+      #
+      # @return [TrueClass,FalseClass] the result of the check
+      #
+      # @since 0.9.2
+      def failure?
         !successful?
       end